backend/requirements.txt
# Framework principal
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-decouple==3.8

# Cache et tâches asynchrones
redis==5.0.1
celery==5.3.4

# Génération de documents
docxtpl==0.16.7
jinja2==3.1.2
openpyxl==3.1.2

# Validation et serialization
pydantic==2.5.2
pydantic[email]==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosqlite==0.19.0
pytest-cov==4.1.0

# Utilitaires
python-dateutil==2.8.2
pytz==2023.3